# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  noises.Request:
    properties:
      noiseLevel:
        maximum: 100
        minimum: 0
        type: integer
      programmingLanguage:
        type: string
      sourceCode:
        type: string
    required:
    - noiseLevel
    - programmingLanguage
    - sourceCode
    type: object
  noises.Response:
    properties:
      responseInfo:
        $ref: '#/definitions/response_info.ResponseInfo'
      taskAlias:
        type: string
    type: object
...
  response_info.ResponseInfo:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  skips.Request:
    properties:
      programmingLanguage:
        type: string
      skipsNumber:
        minimum: 0
        type: integer
      sourceCode:
        type: string
    required:
    - programmingLanguage
    - skipsNumber
    - sourceCode
    type: object
  skips.Response:
    properties:
      responseInfo:
        $ref: '#/definitions/response_info.ResponseInfo'
      taskAlias:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for the Codular project, providing functionality to generate code
    skips, retrieve tasks, and check task status.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Codular API
  version: "1.0"
paths:
...
  /noises/generate:
    post:
      consumes:
      - application/json
      description: Processes the provided source code with a specified noise level,
        generates a unique alias, saves the task with its description to the database,
        and initiates asynchronous processing. Returns the task alias for retrieving
        the task code and description.
      parameters:
      - description: Source code, noise level, and programming language
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/noises.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Example response" Example({"responseInfo":{"status":"OK"},"taskAlias":"abc123"})
          schema:
            $ref: '#/definitions/noises.Response'
        "400":
          description: Invalid request, empty body, or invalid programming language
          schema:
            $ref: '#/definitions/noises.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/noises.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/noises.Response'
      security:
      - Bearer: []
      summary: Generate and save noised code
      tags:
      - Noises
...
  /skips/generate:
    post:
      consumes:
      - application/json
      description: Processes the provided source code with a specified number of skips,
        generates a unique alias, saves the task with its description to the database,
        and initiates asynchronous processing. Returns the task alias for retrieving
        the task code and description.
      parameters:
      - description: Source code, number of skips, and programming language
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/skips.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Example response" Example({"responseInfo":{"status":"OK"},"taskAlias":"abc123"})
          schema:
            $ref: '#/definitions/skips.Response'
        "400":
          description: Invalid request, empty body, or invalid programming language
          schema:
            $ref: '#/definitions/skips.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/skips.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/skips.Response'
      security:
      - Bearer: []
      summary: Generate and save skips code
      tags:
      - Skips
...
  /task/random:
    get:
      description: Redirects to a random public task with public = true. The redirected
        endpoint returns the task code and description.
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to /api/v1/task/{alias}
          schema:
            type: string
        "404":
          description: No public tasks found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get random public task
      tags:
      - Task
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"